Add a tool to verify an r0 chunk with a number of random probes.
This only detects a small fraction of errors, but it's still useful to check
that a file isn't completely incorrect. Since the tool uses random offsets it
can be run multiple times on the same files to obtain a higher degree of
confidence in the correctness of the output files.