Add new solver for even phases: backpropagate2.
This is similar to the existing backpropagate-losses solver, except that
instead of writing wins to a bitmap output file, it encodes newly-discovered
wins using Elias-Fano coding.

The purpose of this change is to make the tool more suitable to distributed
computation because chunk outputs can be calculated independently.

The downside compared to the bitmap approach is that different output chunks
may contain duplicate data. This isn't a problem for correctness but it means
the total output size could become large.

As input, this tool uses two rN bin files from the previous two phases
instead of a bin file and delta file (160 GB instead of 130 GB of input data).
This uses more space but might involve less downloading since participants
can reuse the rN.bin files from the previous phase (so only one 80 GB input
file needs to be downloaded at a time).

Unlike backpropagate-losses, this tool also has support for automatic chunk
assignment from AutomaticSolver.