Add support for reading minimized values from a compressed XZ file.
This allows lookup-min and pushfight-standalone-server to be used with just
the 4 GB minimized.bin.xz file, instead of the uncompressed 86 GB minimized.bin file.

In my experiments this seems to be about 100 times slower, but it still takes only
about 500 ms to analyze a position, so for local play this should be fine.